fix: remove type checks
Due to the infer type circular reference issue,
consumer of `type-plus` needs to turn on `skipLibCheck` in order to use `type-plus`.

It it not a good practice.
Since the type checker needs some more work,
remove it from the package to improve DX.

BREAKING CHANGE the type checker is removed from the package